fix(AssetStore): correct progress calculation in load method to ensure accurate model count

+++ zappy/src/zappy/gui/render/AssetStore.cpp
@@ -26,7 +26,7 @@ void AssetStore::load(const ProgressCallback& onProgress) {
     for (const auto& [name, loadStep] : AssetLoaderRegistry::create(*this)) {
         loadStep();
         if (onProgress) {
-            onProgress(name, static_cast<float>(_playerModels.size() + _resourcesModels.size() + 3.0F) /
+            onProgress(name, (_playerModels.size() + _resourcesModels.size() + 3.0F) /
                                  static_cast<float>(kMaxModels + std::to_underlying(game::ResourceType::Thystame) +
                                                     1.0F + 3.0F));
         }
